Organize cargo zng res built-in tools
